% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clootl_data}
\alias{clootl_data}
\title{A complex data store used in the package.}
\format{
List of csv files and Newick and Nexus phylogenies
}
\source{
\url{https://github.com/eliotmiller/clootl}
}
\usage{
clootl_data
}
\description{
Taxonomy files and phylogenies.
}
\keyword{datasets}
