% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataStore}
\alias{dataStore}
\title{A complex data store used in the package.}
\format{
List of csv files and Newick and Nexus hylogenies
}
\source{
\url{https://github.com/eliotmiller/clootl}
}
\usage{
dataStore
}
\description{
Taxonomy files and phylogenies.
}
\keyword{datasets}
