% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTree.R
\name{extractTree}
\alias{extractTree}
\title{Randomly resolve polytomies and enforce ultrametric}
\usage{
extractTree(species, output.type, taxonomy.year, version, which.tree)
}
\arguments{
\item{species}{A character vector either of scientific names (directly as they come out of the
eBird taxonomy, i.e. without underscores) or of six-letter eBird species codes. Any elements of
the species vector that do not match a species-level taxon in the specified eBird taxonomy
will result in an error. Set to "all.species" if the complete tree is desired.}

\item{output.type}{Either "scientific" or "code".}

\item{taxonomy.year}{The eBird taxonomy year the tree should be output in. Set to "current"
to extract a tree in the most recent taxonomic version.}

\item{version}{The desired version of the tree. Set to current to extract the most recent
version of the tree.}

\item{which.tree}{Not currently implemented, and defaults for now to a placeholder tree. In the
future there will be the base tree from OpenTree, a small cloud of dated, complete trees (which
can be replaced by downloading and plugging in a larger set of such trees using a to-be-written
function for that purpose),
and a maximum clade credibility version of these trees.}
}
\value{
A phylogeny of the specified taxa in the specified eBird taxonomy version and clootl
tree version.
}
\description{
Initial tree processing
}
\details{
This function first ensures that the requested output species overlap with species-level
taxa in the requested eBird taxonomy. If they do not, the function will error out. The onus is
on the user to ensure the requested taxa are valid. This is critical to ensure no unexpected
analysis hiccups later--you don't want to find out many steps later that your dataset doesn't
match your phylogeny.
}
\examples{
ex1 <- extractTree(species=c("amerob", "canwar", "reevir1", "yerwar", "gockin"),
   output.type="code", taxonomy.year="current", version="current")
ex2 <- extractTree(species=c("Turdus migratorius", "Setophaga dominica", "Setophaga ruticilla", "Sitta canadensis"),
   output.type="scientific", taxonomy.year="current", version="current")
}
\author{
Eliot Miller
}
