% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avesdata_repo}
\alias{get_avesdata_repo}
\title{Pull down full AvesData repository to a working directory}
\usage{
get_avesdata_repo(
  url = "https://github.com/McTavishLab/AvesData/archive/refs/heads/main.zip",
  refresh = FALSE
)
}
\arguments{
\item{url}{Web address of the Aves Data repository at https://github.com/McTavishLab/AvesData/}

\item{refresh}{Default to \code{FALSE}. Will not redownload the data by default if path exists, unless refresh=TRUE}
}
\description{
Pull down full AvesData repository to a working directory
}
