% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avesdata_repo_path}
\alias{get_avesdata_repo_path}
\title{Get path to Aves Data folder}
\usage{
get_avesdata_repo_path()
}
\value{
String - path to Aves Data folder, if set. Returns "" if not set.
}
\description{
Get path to Aves Data folder, if set.
}
\details{
Based on https://github.com/CornellLabofOrnithology/auk/blob/main/R/auk-set-ebd-path.r
Use this function to check stored path to downloaded AvesData folder from https://github.com/McTavishLab/AvesData.
When \code{AVESDATA_PATH} is set, the data_path in any clootl functions with a \code{data_path} argument will default to this value.
}
\examples{
\dontrun{
get_avesdata_repo_path()
}
}
