% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_avesdata_repo_path}
\alias{set_avesdata_repo_path}
\title{Set path to Aves Data folder already somewhere on your computer}
\usage{
set_avesdata_repo_path(path)
}
\arguments{
\item{path}{A character vector with the path to the Aves Data folder.}
}
\description{
Set path to Aves Data folder already somewhere on your computer
}
