% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCitations.R
\name{getCitations}
\alias{getCitations}
\title{Get the DOIs and quantify the contribution of published studies}
\usage{
getCitations(tree)
}
\arguments{
\item{tree}{A phylogeny obtained from extractTree (see details).}
}
\value{
A dataframe of the percent of internal nodes supported by a given study, as well
as the DOI of that study. The proportion of taxa in the tree supported by taxonomic
addition only is included in the dataframe.
}
\description{
Standing on the shoulders of giants
}
\details{
Importantly: an internet connection is required for this function to work, as it
relies on Open Tree of Life APIs. The function will determine what proportion of nodes in
your phylogeny (possibly
but not necessarily pruned to a set of study taxa) are supported by each study that goes into
creating the final clootl tree. In any resulting publication, you should always cite the
clootl tree, and
you should also "always" cite all the trees/DOIs that contributed to your phylogeny. That
said, we are well aware of citation and word count limits that plague modern publishing,
and for this reason we quantify the contribution of each study; depending on your phylogeny,
it is very possible that one or two studies contributed the majority of information. Currently,
this function assumes your output tree matches the taxonomy of the corresponding tree on the
OpenTree server. Since the function is actually using the named internal nodes for the API
query, and these should not be lost between tree versions and taxonomies, this should not
matter, but this has not yet been tested.
}
\examples{
\donttest{#pull the taxonomy file out
data(dataStore)
tax <- dataStore$taxonomy.files$year2021

#subset to species only
tax <- tax[tax$CATEGORY=="species",]

#simulate extracting a tree for a particular family
temp <- tax[tax$FAMILY=="Rhinocryptidae (Tapaculos)",]
spp <- temp$SCI_NAME

#get your tree
prunedTree <- extractTree(species=spp, output.type="scientific",
   taxonomy.year=2021, version="current")

#get your citation DF
yourCitations <- getCitations(tree=prunedTree)}
}
\author{
Eliot Miller, Emily Jane McTavish
}
