% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialProcessing.R
\name{initialProcessing}
\alias{initialProcessing}
\title{Randomly resolve polytomies and enforce ultrametric phylogenetic tree}
\usage{
initialProcessing(orig.tree)
}
\arguments{
\item{orig.tree}{Phylogeny in ape format (\code{phylo} object).}
}
\value{
A fully resolved and strictly ultrametric tree.
}
\description{
Initial tree processing
}
\details{
This function relies on a random polytomy resolver re-used, under a GPL-3 license,
from the PDcalc package by David Nipperess (https://github.com/davidnipperess/PDcalc/tree/master).
It also uses code borrowed from a blog post by Jonathan Chang to force a tree to be
strictly ultrametric
(https://www.r-bloggers.com/2021/07/three-ways-to-check-and-fix-ultrametric-phylogenies/).
}
\examples{
#load the base tree and send it through the initial processing steps
data(clootl_data)
# TODO check error next:
# newTree <- initialProcessing(clootl_data$trees$Aves_1.3$summary.trees$year2021)
}
\author{
Eliot Miller, David Nipperess, Jonathan Chang
}
