% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTree.R
\name{treeGet}
\alias{treeGet}
\title{Helper to load a tree into the R environment}
\usage{
treeGet(version, taxonomy_year, data_path = FALSE)
}
\description{
Not exported. Internal use only.
}
\keyword{internal}
